requirements/base.txt
celery==5.2.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.2
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
django-recaptcha==3.0.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environ==1.0
phonenumbers==8.13.9
pika==1.3.2
pillow==9.4.0
pip-chill==1.0.3
pre-commit==3.3.0
psycopg2==2.9.6
pycryptodome==3.17
pyjokes==0.6.0
python-environ==0.4.54
sorl-thumbnail==12.9.0
